consolidate wallstreetbets scraper and store weekly SP500 OHLCV
