+++ docs/data_format.md
@@ -19,7 +19,7 @@ analysis can reproduce past views of the data.
 | `analyst_ratings` | `ticker`, `rating`, `date`, `_retrieved` |
 | `news_headlines` | `ticker`, `headline`, `link`, `source`, `time`, `_retrieved` |
 | `insider_buying` | `ticker`, `exec`, `shares`, `date`, `_retrieved` |
-| `sp500_index` | `date`, `close`, `_retrieved` |
+| `sp500_index` | `date`, `open`, `high`, `low`, `close`, `volume`, `_retrieved` |
 | `ticker_scores` | `symbol`, `index_name`, `date`, `fundamentals`, `momentum`, `liquidity_sentiment`, `risk_adjusted`, `overall` |
 | `top_scores` | `date`, `symbol`, `index_name`, `score`, `rank` |
 | `portfolios` | `id`, `name`, `weights` |
